+++ Destination/src/main/java/com/tcs/destination/helper/NotificationHelper.java
@@ -14,6 +14,7 @@ import static com.tcs.destination.utils.Constants.TOKEN_TO;
 import static com.tcs.destination.utils.Constants.TOKEN_USER;
 
 import java.lang.reflect.InvocationTargetException;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -930,7 +931,7 @@ public class NotificationHelper implements Runnable {
 						opportunity.getOpportunityId());
 		String displayIOU = opportunity.getCustomerMasterT()
 				.getIouCustomerMappingT().getDisplayIou();
-		Integer digitalDealValue = opportunity.getDigitalDealValue();
+		BigDecimal digitalDealValue = opportunity.getDigitalDealValue();
 		String geography = opportunity.getGeographyCountryMappingT()
 				.getGeography();
 		String country = opportunity.getCountry();
