+++ Destination/src/main/java/com/tcs/destination/service/BDMDetailedReportService.java
@@ -2,6 +2,7 @@ package com.tcs.destination.service;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
@@ -401,7 +402,7 @@ public class BDMDetailedReportService {
 				//set project deal value and deal currency
 				if (projectDVFlag) {
 					if(opportunity.getDigitalDealValue() != null){
-						row.createCell(colValue).setCellValue(opportunity.getDigitalDealValue());
+						row.createCell(colValue).setCellValue(opportunity.getDigitalDealValue().doubleValue());
 					} else {
 						row.createCell(colValue).setCellValue(0);
 					}
@@ -672,7 +673,7 @@ public class BDMDetailedReportService {
 				//set project deal value and deal currency
 				if (projectDVFlag) {
 					if(opportunity.getDigitalDealValue() != null){
-						row.createCell(colValue).setCellValue(opportunity.getDigitalDealValue());
+						row.createCell(colValue).setCellValue(opportunity.getDigitalDealValue().doubleValue());
 					} else {
 						row.createCell(colValue).setCellValue(0);
 					}
